+++ test/model/Canvas.js
@@ -1,9 +1,9 @@
 var test = require('tape')
 
-import Canvas from 'abstract/Canvas'
+var Canvas = require('model/Canvas')
 
 test('Abstract Canvas', (t) => {
-  let canvas = new Canvas()
+  var canvas = new Canvas()
 
   t.deepEqual(canvas.link, {}, 'link defaults to {}')
   t.deepEqual(canvas.node, {}, 'node defaults to {}')